Rename TVTextScrollViewExample
